requirements.txt
aiodns==2.0.0
aiohttp==3.7.3
async-timeout==3.0.1
asyncpg==0.21.0
attrs==20.3.0
brotlipy==0.7.0
cchardet==2.1.7
cffi==1.14.4
chardet==3.0.4
idna==2.10
multidict==5.1.0
pycares==3.1.1
pycparser==2.20
python-dotenv==0.15.0
typing-extensions==3.7.4.3
yarl==1.6.3
